Refactoring (tested)
