+++ Development/F7_Validation/L432_Helper/Inc/types.h
@@ -1,0 +1,41 @@
+/**
+  *****************************************************************************
+  * @file    types.h
+  * @author  Tyler
+  *
+  * @defgroup Types
+  * @brief Defines some types shared throughout program
+  * @{
+  *****************************************************************************
+  */
+
+
+
+
+#ifndef TYPES_H
+#define TYPES_H
+
+
+
+
+/********************************* Includes **********************************/
+#include <stdint.h>
+
+
+
+
+/********************************** Types ************************************/
+typedef struct{
+    uint8_t id;
+    uint16_t pos;
+}Data_t;
+
+
+
+
+/**
+ * @}
+ */
+/* end Types */
+
+#endif /* TYPES_H */
